requirements.txt
langchain==1.2.15
langchain-community==0.4.1
langchain-text-splitters==1.1.2
langchain-ollama==1.1.0
pypdf==6.10.2
chromadb==1.5.8
python-dotenv==1.2.2
fastapi==0.136.0
uvicorn==0.45.0
pydantic==2.13.3
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.24.0
requests==2.33.1
